Add membership validations (#75)
* Refactor: Use let in spec

This allows to scope the instatntiation so that they do not persist
after the test block runs.

* Validate organization presence unless superuser

* Validates unique membership row

Takes into consideration the role and user columns. This is to ensure a
user is not assigned the same role twice. In the future this could very
well change if users are allowed to a member of more than one
organization.

* Superuser cannot belong to an organization

Errors on validation if trying to assign an organization to a superuser.

* Cache superuser role id

* Replace create method with new

* Admin cannot be member and vice versa

This is done so that admin/member organization can be easily found as
there will be only one matching row.

* Do not allow user to be member of another organization