Change user/organizations relationship to has_many
This will allow for querying the database to easily see which
organization the user belongs to. This is needed since an organization
can have many users; otherwise the through option on the has_one will
not work. The has one association constraint can still be enforced which
it currently is through the membership model.