Move user registration to admin page (#76)
* Remove registerable from user model

This allows us to control who has the ability to add a user.

* Add sign in link to application view

* Add user create action/views

This is being done because only the admin will have the ability to add
users.

* Add 'Add User' link to admin page

* Create policy for new/create actions

This grants superuser access for new and create actions.

* Grant admin access to new/create action

* Create index action policy

* Create user model methods admin/admin_for/superuser

* Change user/organizations relationship to has_many

This will allow for querying the database to easily see which
organization the user belongs to. This is needed since an organization
can have many users; otherwise the through option on the has_one will
not work. The has one association constraint can still be enforced which
it currently is through the membership model.

* Show action has unrestricted access

* Create policy for destroy action

This allows a superuser or an organizations admin.

* User roles query

* Move id out of superuser_role method

The memberships method superuser_role should return the superuser_role
not the id as that is what its method name implies.

* Deny access for admin to index users

This will be created in separate issue as an admin can only index users
belonging to their organization.

* Configure sign_in method for request spec

* Add authorization checks to user controller

* Create initial request spec